+++ AcademyPlan/src/entities/Aluno.java
@@ -12,11 +12,11 @@ public class Aluno extends Pessoa {
     private PlanoAlimentar planoAlimentar;
     private Treino treino;
 
-    public Aluno(String name, LocalDate dataNascimento, Integer telefone, Integer cpf, Integer cep) {
+    public Aluno(String name, LocalDate dataNascimento, Long telefone, Long cpf, Integer cep) {
     	 super(name, dataNascimento, telefone, cpf, cep);
     }
 
-    public Aluno(String name, LocalDate dataNascimento, Integer telefone, Integer cpf, Integer cep,
+    public Aluno(String name, LocalDate dataNascimento, Long telefone, Long cpf, Integer cep,
             PlanoAcademia planoAcademia, PlanoAlimentar planoAlimentar, Treino treino) {
         super(name, dataNascimento, telefone, cpf, cep);
         this.planoAcademia = planoAcademia;
